requirements.txt
boto3==1.24.81
dill==0.3.5.1
dm-haiku==0.0.7
jax==0.4.6
jaxlib==0.4.6
matplotlib==3.5.3
mplcyberpunk==0.6.0
numpy==1.23.2
opencv-python==4.6.0.66
pandas==1.5.1
Pillow==9.2.0
PyYAML==6.0
requests==2.28.1
scipy==1.9.0
scikit-learn==1.1.3
seaborn==0.12.1
torch==1.12.1
torchvision==0.13.1
torchdata==0.4.1
tensorboard==2.10.1
